refactor: rename timers as per their function, auto refresh timer now every 15 seconds and only after we confirm bitcoind is repsonsive
Only query bitcoind status with a `pgrep` script as using rpc commands for this is not great.  More informative status colors for Tor and Bitcoin, white means we do not yet know their status, orange means its running but we are not sure if its booted completely or that we are in process of shutting down/booting up.